All classes with member var "fn" now have member var "name"
This is an optional member var which can be manually set by the user.
If not, in the first instance it will be replaced with name of variable passed to "fn".
If a literal is passed to "fn", instead the index of the hazard/transition/trajectory is strigified.

This name will be used to identify the relevant object, when reporting warnings/errors related to return values from user functions.